NIFI-13507 removed nifi prometheus bundle, moved prometheus extension utils to web api.  Removed StandardGangliaReporter
