+++ nifi-assembly/pom.xml
@@ -420,12 +420,6 @@ language governing permissions and limitations under the License. -->
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-prometheus-nar</artifactId>
-            <version>2.0.0-SNAPSHOT</version>
-            <type>nar</type>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.nifi</groupId>
             <artifactId>nifi-network-processors-nar</artifactId>
             <version>2.0.0-SNAPSHOT</version>
             <type>nar</type>
