+++ nifi-code-coverage/pom.xml
@@ -892,11 +892,6 @@
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-prometheus-utils</artifactId>
-            <version>2.0.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.nifi</groupId>
             <artifactId>nifi-put-pattern</artifactId>
             <version>2.0.0-SNAPSHOT</version>
         </dependency>
@@ -1221,11 +1216,6 @@
         <dependency>
             <groupId>org.apache.nifi</groupId>
             <artifactId>nifi-poi-services</artifactId>
-            <version>2.0.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-prometheus-reporting-task</artifactId>
             <version>2.0.0-SNAPSHOT</version>
         </dependency>
         <dependency>
