--- nifi-extension-bundles/nifi-prometheus-bundle/nifi-prometheus-nar/src/main/resources/META-INF/NOTICE
@@ -1,76 +1,0 @@
-Copyright 2015-2017 The Apache Software Foundation
-
-This product includes software developed at
-The Apache Software Foundation (http://www.apache.org/).
-
-******************
-Apache Software License v2
-******************
-
-The following binary components are provided under the Apache Software License v2
-
-  (ASLv2) Dropwizard Metrics
-    The following NOTICE information applies:
-      Metrics
-      Copyright 2010-2013 Coda Hale and Yammer, Inc., 2014-2017 Dropwizard Team
-      This product includes software developed by Coda Hale and Yammer, Inc.
-
-      This product includes code derived from the JSR-166 project (ThreadLocalRandom, Striped64,
-      LongAdder), which was released with the following comments:
-
-          Written by Doug Lea with assistance from members of JCP JSR-166
-          Expert Group and released to the public domain, as explained at
-          http://creativecommons.org/publicdomain/zero/1.0/
-
-      The derived work in the nifi-metrics module is adapted from
-        https://github.com/dropwizard/metrics/blob/v2.2.0/metrics-core/src/main/java/com/yammer/metrics/core/VirtualMachineMetrics.java
-      and can be found in
-        nifi-commons/nifi-metrics/src/main/java/org/apache/nifi/metrics/jvm/JvmMetrics.java
-        nifi-commons/nifi-metrics/src/main/java/org/apache/nifi/metrics/jvm/JmxJvmMetrics.java
-
-  (ASLv2) Prometheus Simple Client libraries
-    Copyright 2012-2019 The Prometheus Authors
-
-  (ASLv2) Jetty
-    The following NOTICE information applies:
-       Jetty Web Container
-       Copyright 1995-2017 Mort Bay Consulting Pty Ltd.
-
-  (ASLv2) Apache Commons Codec
-      The following NOTICE information applies:
-        Apache Commons Codec
-        Copyright 2002-2014 The Apache Software Foundation
-
-        src/test/org/apache/commons/codec/language/DoubleMetaphoneTest.java
-        contains test data from http://aspell.net/test/orig/batch0.tab.
-        Copyright (C) 2002 Kevin Atkinson (kevina@gnu.org)
-
-        ===============================================================================
-
-        The content of package org.apache.commons.codec.language.bm has been translated
-        from the original php source code available at http://stevemorse.org/phoneticinfo.htm
-        with permission from the original authors.
-        Original source copyright:
-        Copyright (c) 2008 Alexander Beider & Stephen P. Morse.
-
-    (ASLv2) Apache Commons Lang
-    The following NOTICE information applies:
-      Apache Commons Lang
-      Copyright 2001-2017 The Apache Software Foundation
-
-      This product includes software from the Spring Framework,
-      under the Apache License 2.0 (see: StringUtils.containsWhitespace())
-	
-
-************
-MIT
-************
-
-  	This product optionally depends on 'Bouncy Castle Crypto APIs' to generate
-  	a temporary self-signed X.509 certificate when the JVM does not provide the
-  	equivalent functionality.  It can be obtained at:
-
-  	  * LICENSE:
-  	    * license/LICENSE.bouncycastle.txt (MIT License)
-  	  * HOMEPAGE:
-  	    * http://www.bouncycastle.org/
