+++ nifi-extension-bundles/pom.xml
@@ -75,7 +75,6 @@
         <module>nifi-extension-utils</module>
         <module>nifi-redis-bundle</module>
         <module>nifi-network-bundle</module>
-        <module>nifi-prometheus-bundle</module>
         <module>nifi-sql-reporting-bundle</module>
         <module>nifi-hazelcast-bundle</module>
         <module>nifi-asn1-bundle</module>
