+++ nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-api/pom.xml
@@ -459,11 +459,6 @@
             <scope>provided</scope> <!-- expected to be provided by parent classloader -->
         </dependency>
         <dependency>
-            <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-prometheus-utils</artifactId>
-            <version>2.0.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
             <groupId>io.swagger.core.v3</groupId>
             <artifactId>swagger-annotations</artifactId>
             <scope>provided</scope> <!-- expected to be provided by parent classloader -->
