+++ nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-api/src/main/java/org/apache/nifi/prometheusutil/PrometheusMetricsUtil.java
@@ -15,11 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.nifi.prometheus.util;
+package org.apache.nifi.prometheusutil;
 
 import io.prometheus.client.CollectorRegistry;
 import org.apache.nifi.components.AllowableValue;
-import org.apache.nifi.components.PropertyDescriptor;
 import org.apache.nifi.controller.status.ConnectionStatus;
 import org.apache.nifi.controller.status.PortStatus;
 import org.apache.nifi.controller.status.ProcessGroupStatus;
@@ -28,10 +27,8 @@ import org.apache.nifi.controller.status.RemoteProcessGroupStatus;
 import org.apache.nifi.controller.status.TransmissionStatus;
 import org.apache.nifi.controller.status.analytics.StatusAnalytics;
 import org.apache.nifi.diagnostics.StorageUsage;
-import org.apache.nifi.expression.ExpressionLanguageScope;
 import org.apache.nifi.metrics.jvm.JvmMetrics;
 import org.apache.nifi.processor.DataUnit;
-import org.apache.nifi.processor.util.StandardValidators;
 import org.apache.nifi.util.StringUtils;
 
 import java.util.Map;
@@ -39,58 +36,14 @@ import java.util.concurrent.TimeUnit;
 
 public class PrometheusMetricsUtil {
 
-    public static final AllowableValue METRICS_STRATEGY_ROOT = new AllowableValue("Root Process Group", "Root Process Group",
-            "Send rollup metrics for the entire root process group");
     public static final AllowableValue METRICS_STRATEGY_PG = new AllowableValue("All Process Groups", "All Process Groups",
             "Send metrics for each process group");
     public static final AllowableValue METRICS_STRATEGY_COMPONENTS = new AllowableValue("All Components", "All Components",
             "Send metrics for each component in the system, to include processors, connections, controller services, etc.");
 
-    private static final CollectorRegistry CONNECTION_ANALYTICS_REGISTRY = new CollectorRegistry();
-    private static final CollectorRegistry BULLETIN_REGISTRY = new CollectorRegistry();
-
     protected static final String DEFAULT_LABEL_STRING = "";
     private static final double MAXIMUM_BACKPRESSURE = 1.0;
     private static final double UNDEFINED_BACKPRESSURE = -1.0;
-
-    // Common properties/values
-    public static final AllowableValue CLIENT_NONE = new AllowableValue("No Authentication", "No Authentication",
-            "ReportingTask will not authenticate clients. Anyone can communicate with this ReportingTask anonymously");
-    public static final AllowableValue CLIENT_WANT = new AllowableValue("Want Authentication", "Want Authentication",
-            "ReportingTask will try to verify the client but if unable to verify will allow the client to communicate anonymously");
-    public static final AllowableValue CLIENT_NEED = new AllowableValue("Need Authentication", "Need Authentication",
-            "ReportingTask will reject communications from any client unless the client provides a certificate that is trusted by the TrustStore"
-                    + "specified in the SSL Context Service");
-
-    public static final PropertyDescriptor METRICS_ENDPOINT_PORT = new PropertyDescriptor.Builder()
-            .name("prometheus-reporting-task-metrics-endpoint-port")
-            .displayName("Prometheus Metrics Endpoint Port")
-            .description("The Port where prometheus metrics can be accessed")
-            .required(true)
-            .expressionLanguageSupported(ExpressionLanguageScope.ENVIRONMENT)
-            .defaultValue("9092")
-            .addValidator(StandardValidators.INTEGER_VALIDATOR)
-            .build();
-
-    public static final PropertyDescriptor INSTANCE_ID = new PropertyDescriptor.Builder()
-            .name("prometheus-reporting-task-instance-id")
-            .displayName("Instance ID")
-            .description("Id of this NiFi instance to be included in the metrics sent to Prometheus")
-            .required(true)
-            .expressionLanguageSupported(ExpressionLanguageScope.ENVIRONMENT)
-            .defaultValue("${hostname(true)}")
-            .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
-            .build();
-
-    public static final PropertyDescriptor CLIENT_AUTH = new PropertyDescriptor.Builder()
-            .name("prometheus-reporting-task-client-auth")
-            .displayName("Client Authentication")
-            .description("Specifies whether or not the Reporting Task should authenticate clients. This value is ignored if the <SSL Context Service> "
-                    + "Property is not specified or the SSL Context provided uses only a KeyStore and not a TrustStore.")
-            .required(true)
-            .allowableValues(CLIENT_NONE, CLIENT_WANT, CLIENT_NEED)
-            .defaultValue(CLIENT_NONE.getValue())
-            .build();
 
     public static CollectorRegistry createNifiMetrics(NiFiMetricsRegistry nifiMetricsRegistry, ProcessGroupStatus status,
                                                       String instId, String parentProcessGroupId, String compType, String metricsStrategy) {
