+++ nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-api/src/main/java/org/apache/nifi/web/util/VirtualThreadParallelProcessingService.java
@@ -19,8 +19,8 @@ package org.apache.nifi.web.util;
 import jakarta.ws.rs.WebApplicationException;
 import org.apache.nifi.connectable.Connection;
 import org.apache.nifi.controller.status.analytics.StatusAnalytics;
-import org.apache.nifi.prometheus.util.ConnectionAnalyticsMetricsRegistry;
-import org.apache.nifi.prometheus.util.PrometheusMetricsUtil;
+import org.apache.nifi.prometheusutil.ConnectionAnalyticsMetricsRegistry;
+import org.apache.nifi.prometheusutil.PrometheusMetricsUtil;
 import org.apache.nifi.util.FormatUtils;
 import org.apache.nifi.util.NiFiProperties;
 import org.apache.nifi.web.controller.ControllerFacade;
