+++ nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-api/src/test/java/org/apache/nifi/web/api/TestFlowResource.java
@@ -27,12 +27,12 @@ import io.prometheus.client.Collector.MetricFamilySamples.Sample;
 import io.prometheus.client.CollectorRegistry;
 import io.prometheus.client.exporter.common.TextFormat;
 import org.apache.nifi.metrics.jvm.JmxJvmMetrics;
-import org.apache.nifi.prometheus.util.BulletinMetricsRegistry;
-import org.apache.nifi.prometheus.util.ClusterMetricsRegistry;
-import org.apache.nifi.prometheus.util.ConnectionAnalyticsMetricsRegistry;
-import org.apache.nifi.prometheus.util.JvmMetricsRegistry;
-import org.apache.nifi.prometheus.util.NiFiMetricsRegistry;
-import org.apache.nifi.prometheus.util.PrometheusMetricsUtil;
+import org.apache.nifi.prometheusutil.BulletinMetricsRegistry;
+import org.apache.nifi.prometheusutil.ClusterMetricsRegistry;
+import org.apache.nifi.prometheusutil.ConnectionAnalyticsMetricsRegistry;
+import org.apache.nifi.prometheusutil.JvmMetricsRegistry;
+import org.apache.nifi.prometheusutil.NiFiMetricsRegistry;
+import org.apache.nifi.prometheusutil.PrometheusMetricsUtil;
 import org.apache.nifi.registry.flow.FlowVersionLocation;
 import org.apache.nifi.web.NiFiServiceFacade;
 import org.apache.nifi.web.ResourceNotFoundException;
